Refactor signed in header
