+++ front/app/containers/HomePage/SignedInHeader/CompleteProfileStep.tsx
@@ -13,9 +13,15 @@ import messages from '../messages';
 
 import AcceptButton from './AcceptButton';
 import OnboardingStep from './OnboardingStep';
+import {
+  Left,
+  Right,
+  Text,
+  HeaderContent,
+  Icons,
+  StyledAvatar,
+} from './Shared';
 import SkipButton from './SkipButton';
-
-import { Left, Right, Text, HeaderContent, Icons, StyledAvatar } from './';
 
 const CompleteProfileIcon = styled(Icon)`
   width: 48px;
