+++ front/app/containers/HomePage/SignedInHeader/index.tsx
@@ -1,6 +1,6 @@
 import React, { lazy, Suspense } from 'react';
 
-import { media, fontSizes, isRtl } from '@citizenlab/cl2-component-library';
+import { Box, media } from '@citizenlab/cl2-component-library';
 import styled from 'styled-components';
 
 import { OnboardingCampaignName } from 'api/onboarding_campaigns/types';
@@ -9,14 +9,16 @@ import useDismissOnboardingCampaign from 'api/onboarding_campaigns/useDismissOnb
 
 import { IHomepageBannerSettings } from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/CraftComponents/HomepageBanner';
 
-import Avatar from 'components/Avatar';
+import {
+  Container,
+  HeaderImage,
+  HeaderImageBackground,
+  HeaderImageOverlay,
+} from 'components/LandingPages/citizen/FullWidthBannerLayout';
 
 import { trackEventByName } from 'utils/analytics';
-import { isNilOrError } from 'utils/helperUtils';
 
 import tracks from '../tracks';
-
-import HeaderImage from './HeaderImage';
 
 const CompleteProfileStep = lazy(() => import('./CompleteProfileStep'));
 const VerificationOnboardingStep = lazy(
@@ -25,118 +27,29 @@ const VerificationOnboardingStep = lazy(
 const CustomCTAStep = lazy(() => import('./CustomCTAStep'));
 const FallbackStep = lazy(() => import('./FallbackStep'));
 
+const heights = {
+  phone: 300,
+  tablet: 250,
+  desktop: 200,
+};
+
 const Header = styled.div`
   width: 100%;
-  height: 190px;
+  height: ${heights.desktop}px;
+  margin: 0;
+  padding: 0;
   position: relative;
   display: flex;
   flex-direction: column;
-
-  ${media.tablet`
-    height: auto;
-    max-height: 320px;
-  `}
-`;
-
-export const HeaderContent = styled.div`
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  display: flex;
   align-items: center;
-  justify-content: space-between;
-  padding-top: 20px;
-  padding-bottom: 20px;
-  padding-left: 75px;
-  padding-right: 75px;
-  overflow: hidden;
-
-  ${isRtl`
-    flex-direction: row-reverse;
-  `}
-
-  h2 {
-    color: #fff;
-    font-size: ${fontSizes.xxl}px;
-    line-height: 33px;
-    font-weight: 400;
-  }
+  justify-content: center;
 
   ${media.tablet`
-    padding-left: 30px;
-    padding-right: 30px;
-  `}
-
-  ${media.tablet`
-    flex-direction: column;
-    align-items: stretch;
-    justify-content: center;
-    padding-left: 15px;
-    padding-right: 15px;
-  `}
-`;
-
-export const Left = styled.div`
-  display: flex;
-  align-items: center;
-  margin-right: 60px;
-
-  ${isRtl`
-    flex-direction: row-reverse;
-  `}
-
-  ${media.tablet`
-    margin-right: 0;
-    flex-direction: column;
-    align-items: flex-start;
+    height: ${heights.tablet}px;
   `}
 
   ${media.phone`
-    align-items: center;
-    margin-bottom: 30px;
-  `}
-`;
-
-export const Text = styled.div`
-  ${isRtl`
-    direction: rtl;
-  `}
-
-  ${media.phone`
-    text-align: center;
-  `}
-`;
-
-export const Right = styled.div`
-  display: flex;
-
-  ${isRtl`
-    flex-direction: row-reverse;
-  `}
-
-  ${media.phone`
-    flex-direction: column;
-  `}
-`;
-
-export const StyledAvatar = styled(Avatar)`
-  margin-right: -3px;
-  z-index: 2;
-`;
-
-export const Icons = styled.div`
-  display: flex;
-  margin-right: 30px;
-
-  ${isRtl`
-    margin-right: 0px;
-    margin-left: 30px;
-  `}
-
-  ${media.phone`
-    margin-right: 0;
+    height: ${heights.phone}px;
   `}
 `;
 
@@ -157,41 +70,67 @@ const SignedInHeader = ({
     dismissOnboardingCampaign(name);
   };
 
-  if (!isNilOrError(currentOnboardingCampaign)) {
-    const onboardingCampaignName = isContentBuilderDisplay
-      ? 'default'
-      : currentOnboardingCampaign.data.attributes.name;
+  if (!currentOnboardingCampaign) return null;
 
-    return (
-      <Header className={`e2e-signed-in-header`} id="hook-header">
-        <HeaderImage homepageSettings={homepageSettings} />
-        <Suspense fallback={null}>
-          <VerificationOnboardingStep
-            currentOnboardingCampaignName={onboardingCampaignName}
-            onSkip={handleSkip('verification')}
+  const onboardingCampaignName = isContentBuilderDisplay
+    ? 'default'
+    : currentOnboardingCampaign.data.attributes.name;
+
+  const homepageSettingColor =
+    homepageSettings.banner_signed_in_header_overlay_color;
+  const homepageSettingOpacity =
+    homepageSettings.banner_signed_in_header_overlay_opacity;
+
+  return (
+    <Container
+      data-testid="full-width-banner-layout"
+      data-cy="e2e-full-width-banner-layout-container"
+      className={`e2e-signed-out-header`}
+    >
+      <Header id="hook-header">
+        <HeaderImage id="hook-header-image">
+          <HeaderImageBackground
+            data-testid="full-width-banner-layout-header-image"
+            data-cy="e2e-full-width-banner-layout-header-image"
+            src={homepageSettings.header_bg?.large || null}
           />
-          <CompleteProfileStep
-            currentOnboardingCampaignName={onboardingCampaignName}
-            onSkip={handleSkip('complete_profile')}
-          />
-          {/*
+          {homepageSettingColor &&
+            typeof homepageSettingOpacity === 'number' && (
+              <HeaderImageOverlay
+                data-cy="e2e-full-width-layout-header-image-overlay"
+                overlayColor={homepageSettingColor}
+                overlayOpacity={homepageSettingOpacity}
+              />
+            )}
+        </HeaderImage>
+
+        <Box>
+          <Suspense fallback={null}>
+            <VerificationOnboardingStep
+              currentOnboardingCampaignName={onboardingCampaignName}
+              onSkip={handleSkip('verification')}
+            />
+            <CompleteProfileStep
+              currentOnboardingCampaignName={onboardingCampaignName}
+              onSkip={handleSkip('complete_profile')}
+            />
+            {/*
             This step is configured via AdminHQ.
-            See https://citizenlab.atlassian.net/browse/CL-2289
-          */}
-          <CustomCTAStep
-            currentOnboardingCampaignName={onboardingCampaignName}
-            onSkip={handleSkip('custom_cta')}
-          />
-          <FallbackStep
-            currentOnboardingCampaignName={onboardingCampaignName}
-            homepageSettings={homepageSettings}
-          />
-        </Suspense>
+              See https://citizenlab.atlassian.net/browse/CL-2289
+            */}
+            <CustomCTAStep
+              currentOnboardingCampaignName={onboardingCampaignName}
+              onSkip={handleSkip('custom_cta')}
+            />
+            <FallbackStep
+              currentOnboardingCampaignName={onboardingCampaignName}
+              homepageSettings={homepageSettings}
+            />
+          </Suspense>
+        </Box>
       </Header>
-    );
-  }
-
-  return null;
+    </Container>
+  );
 };
 
 export default SignedInHeader;
